% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymp_var.R
\name{asymp_var}
\alias{asymp_var}
\title{Asymptotic variacne function}
\usage{
asymp_var(
  estimationMethod,
  d,
  alphaDash = NA,
  sigma = NA,
  sigma_0_squared = NA,
  SG = NA,
  indexset = NA
)
}
\arguments{
\item{estimationMethod}{If \code{"OracleSigma"} is chosen, the funciton calculates the asymptotic variance for the oracle-volatility estimator. Therefore, provide the true parameters \code{alphaDah,sigma} of the model.
If \code{"SigmaAndKappa"} is chosen, please provide the true parameters \code{alphaDash, sigma_0_squared,indexset,SG}. Note that in this case, the function returns the approximated asymptotic variance, dependent on the provided spatial coordinates.
For additional information, see \link{estimateParametersMulti}.}

\item{d}{Spatial dimension of the SPDE model.}
}
\value{
a value/matrix containing the asymptotic variance of the provided estimation method.
}
\description{
Generates the asymptotic variance of the parameter estimators in the multi-dimensional SPDE model.
}
\examples{

d <- 2
alphaDash <- 0.5
sigma_0_squared <- 1
indexset <- c(15,47,83)
SG <- do.call(CJ,replicate(d,seq(0,1,1/10),F))
av <- asymp_var("SigmaAndKappa",d,alphaDash=alphaDash,sigma_0_squared=sigma_0_squared,indexset = indexset,SG = SG)
}
\seealso{
\link{SecondOrderSPDEMulti},\link{estimateParametersMulti}.
}
\keyword{asymptotic}
\keyword{variance}
