% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParametersMulti.R
\name{estimateParametersSPDEMulti}
\alias{estimateParametersSPDEMulti}
\title{Estimate the parameters of a second-order SPDE model in multiple space dimensions}
\usage{
estimateParametersSPDEMulti(
  data_list,
  estimationMethod,
  kappa = NA,
  eta = NA,
  alphaDash = NA,
  spatialDelta = 0.05,
  indexset = NA,
  ignoreWarnings = F
)
}
\arguments{
\item{data_list}{return of the function 'simulateSPDEmodelMulti' or a list containing multiple returns of the function 'simulateSPDEmodelMulti'.}

\item{estimationMethod}{a string indicating the parameter/parameters to be estimated. If only 'sigma' is unknown choose \code{"OracleSigma"} and provide \code{delta,alphaDash,kappa,eta}, respectively.
If 'sigma' and 'kappa' are known choose \code{"SigmaAndKappa"} and provide the known parameter \code{alphaDash} as well as a \code{indexset} satisfying the respective Assumptions.
If none of the parameters are known, choose \code{"alphaDash"} or \code{"all"}. For \code{"alphaDash"} provide \code{spatialDelta} and for \code{"all"} provide \code{spatialDelta} and \code{indexset}.}

\item{spatialDelta}{a real number greater than zero and less than 1/2 for selecting only the data points which are delta away from the Dirichlet boundary condition. The default is 0.05.}

\item{ignoreWarnings}{if True, the statistical warnings are suppressed. Default is False.}

\item{...}{further arguments depending on the chosen estimation method. See \code{estiamtionmethod} or the examples below.}
}
\value{
a named numeric vector denoting the respective estimate. For method \code{"OracleSigma"} the returned value denotes the estimation of 'sigma^2'. For method \code{"SigmaAndKappa"} the returned values denote the estimate of 'sigma^2' and the estimated quotient 'theta1/theta2' of the respective axis.
For the method \code{"alphaDash"} returns the estimate for 'alphaDash'. \code{"all"} first estimates 'alphaDash' and based on this estimate 'sigma0^2' and each 'kappa'.
See Bossert, P. (2023) for details on estimation methods.
}
\description{
Oracle and non-oracle estimations for SPDE models. All estimators are consistent and estimate the natural parameters of the d-dimensional SPDE model. Non-oracle estimations for second-order SPDEs are based on a log-linear model. For further information, see reference Bossert, P. (2023).
}
\examples{
d <- 2
N <- 10000
M <- 10
theta0 <- 0
eta <- 1
nu <- c(2,1)
sigma <- 1
alphaDash <- 0.5
L <- 20
K <- 50

res <- simulateSPDEmodelMulti(d=d,theta0=theta0,nu=nu,eta=eta,sigma=sigma,
alphaDash=alphaDash,numberOfSpatialPoints=M,
numberOfTemporalPoints=N,L=L,K=K)

estimateParametersSPDEMulti(res,estimationMethod="OracleSigma",spatialDelta=0.05,alphaDash=alphaDash,kappa=nu/eta,eta=eta)
estimateParametersSPDEMulti(res,estimationMethod="SigmaAndKappa",alphaDash=alphaDash, indexset = c(16,53,57))
estimateParametersSPDEMulti(res,estimationMethod="alphaDash",spatialDelta=0.05)
estimateParametersSPDEMulti(res,estimationMethod="all",spatialDelta=0.05,indexset = c(16,53,57))
}
\references{
Bossert, P. (2023), 'Parameter estimation for second-order SPDEs in multiple space dimensions'.
}
\seealso{
\link{simulateSPDEmodelMulti}, \link{MCSPDESamplesMulti},\link{SecondOrderSPDEMulti},
}
\keyword{Estimation}
\keyword{Parameter}
\keyword{SPDEs.}
\keyword{for}
