% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_SP.R
\name{get_filtered_SP}
\alias{get_filtered_SP}
\title{Spatial grid filter}
\usage{
get_filtered_SP(SG, spatialDelta)
}
\arguments{
\item{SG}{a data frame with the multi-dimensional spatial grid coordinates. Such a data-frame is also provided by the function \link{simulateSPDEmodelMulti}.}

\item{spatialDelta}{filter variable. Each entry has to be greater or equal \code{spatialDelta} and less or equal 1-\code{spatialDelta}.}
}
\value{
a data frame containing the filtered spatial grid.
}
\description{
Filters the spatial grid by \code{delta}, where each entry has to be greater or equal \code{spatialDelta} and less or equal 1-\code{spatialDelta} and returns filtered data-frame.
}
\examples{
library(data.table)
M <-10
d <- 3
SG <- do.call(CJ,replicate(d,seq(0,1,1/M),F))
delta <- 0.05

get_filtered_SP(SG,delta)
}
\seealso{
\link{SecondOrderSPDEMulti}, \link{simulateSPDEmodelMulti}
}
\keyword{Filtered}
\keyword{grid}
\keyword{satial}
