% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_SP_index.R
\name{get_filtered_SP_index}
\alias{get_filtered_SP_index}
\title{Spatial grid filter index}
\usage{
get_filtered_SP_index(SG, spatialDelta)
}
\arguments{
\item{SG}{a data frame with the multi-dimensional spatial grid coordinates.}

\item{spatialDelta}{filter variable. Each entry has to be greater or equal \code{spatialDelta} and less or equal 1-\code{spatialDelta}.}
}
\value{
a vector containing the indexset of the spatial grid coordinates which satisfies the 'delta' condition.
}
\description{
Filters the spatial grid by 'delta', where each entry has to be greater or equal \code{spatialDelta} and less or equal 1-\code{spatialDelta} and returns index of the original spatial grid data frame.
}
\examples{
library(data.table)
M <-10
d <- 3
SG <- do.call(CJ,replicate(d,seq(0,1,1/M),F))
delta <- 0.05

get_filtered_SP_index(SG,delta)
}
\seealso{
\link{SecondOrderSPDEMulti}
}
\keyword{filtered}
\keyword{for}
\keyword{grid}
\keyword{index}
\keyword{spatial}
