% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSPDESamplesMulti.R
\name{MCSPDESamplesMulti}
\alias{MCSPDESamplesMulti}
\title{Simulates multiple second-order stochastic partial differential equations in multiple space dimension using Monte-Carlo repetitions}
\usage{
MCSPDESamplesMulti(
  repetitions,
  saveData = F,
  start = 1,
  save_path = "",
  d,
  theta0,
  nu,
  eta,
  sigma,
  alphaDash,
  numberOfSpatialPoints,
  numberOfTemporalPoints,
  L,
  K = NA,
  em_data = NA,
  approx_var = NA,
  save_approx_var = F,
  path_approx_var = "",
  save_em_data = F,
  path_em_data = "",
  mehtod = "replacement",
  numCores = NA
)
}
\arguments{
\item{repetitions}{A natural number giving the number of the Monte-Carlo repetitions.}

\item{...}{for further arguments see \link{simulateSPDEmodelMulti}.}
}
\value{
A list of the returns of \link{simulateSPDEmodelMulti}.
}
\description{
Simulates a Monte-Carlo study of a SPDE model in multiple space dimension on a discrete grid with \code{N} temporal and \code{M} spatial points, where the grid points are equidistant within the unit square. Each repetition uses the "replacement"-method by default, see \link{simulateSPDEmodelMulti}.
}
\examples{
reps <- 10
d <- 2
N <- 1000
M <- 10
theta0 <- 0
eta <- 1
nu <- c(2,1)
sigma <- 1
alphaDash <- 0.5
L <- 20
K <- 50
va <- variance_approx(d,theta0,nu,eta,sigma,alphaDash,M,L,K)

MCSPDESamplesMulti(repetitions=reps,d=d,theta0=theta0,nu=nu,eta=eta,sigma=sigma,alphaDash=alphaDash,numberOfSpatialPoints=M,numberOfTemporalPoints=N,L=L,approx_var=va)
}
\references{
Bossert, P. (2023), 'Parameter estimation for second-order SPDEs in multiple space dimensions'
}
\seealso{
\link{simulateSPDEmodelMulti},\link{SecondOrderSPDEMulti}.
}
\keyword{Monte-Calro}
\keyword{SPDE.}
\keyword{samples}
