% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SPDEMulti.R
\name{plot_SPDEMulti}
\alias{plot_SPDEMulti}
\title{Plot a second-order stochastic partial differential equations in multiple space dimension}
\usage{
plot_SPDEMulti(
  data_list,
  coord_plot,
  spatialCoordsRemainingAxes = NA,
  scale = 1.5,
  camerPosition = c(1.3, 1.3, 1.3/2.5)
)
}
\arguments{
\item{data_list}{the return of the the SPDE simulation, see \link[ParabolicSPDEsMulti:simulateSPDEmodelMulti]{ParabolicSPDEsMulti::simulateSPDEmodelMulti}.}

\item{coord_plot}{a integer in the range of 1 to d, indicating the respective axis to be ploted.}

\item{spatialCoordsRemainingAxes}{a optional (d-1)-dimensional vector with spatial coordinates on the other axis to be fixed. For instance: 3-dimensions (d=3) and we want to plot the 2. axis, where we fix the 1. axis at 0.3 and the 3. axis at 0.7. Then, 'spatialCoordsRemainingAxes'=c(0.3,0.7). If 'spatialCoordsRemainingAxes'=NA (default), then the mid point on each remaining axis is selected. Note, that the chosen values for the remaining axes have to be on the simulated grid.}

\item{scale}{a optional parameter indicating the zoom of the camera position.}

\item{cameraPosition}{an optional 3-dimensional parameter indicating the camera position. See 'plotly' documentation for further information.}
}
\value{
a plotly plot of the generated SPDE model.
}
\description{
Plots the generated multidimensional SPDE data using the 'plotly' package, generated by \link[ParabolicSPDEsMulti:simulateSPDEmodelMulti]{ParabolicSPDEsMulti::simulateSPDEmodelMulti}.
}
\examples{
d <- 3
N <- 50
M <- 20
theta0 <- 0
eta <- 1
nu <- c(4,-0.5,0)
sigma <- 1
alphaDash <- 0.5
L <- 10

res <- simulateSPDEmodelMulti(d=d,theta0=theta0,nu=nu,eta=eta,sigma=sigma,
alphaDash=alphaDash,numberOfSpatialPoints=M,numberOfTemporalPoints=N,L=L)

SPDE_plot(data_list = res, coord_plot = 2, spatialCoordsRemainingAxes = c(0.3,0.7))
}
\seealso{
\link{simulateSPDEmodelMulti},\link{SecondOrderSPDEMulti}
}
