% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_plot.R
\name{RV_plot}
\alias{RV_plot}
\title{Plot realized volatility against the theoretical result}
\usage{
RV_plot(data_list, spatialDelta = 0.05, newColors = T)
}
\arguments{
\item{data_list}{return of the function \code{simulateSPDEmodelMulti} or a list containing the returns of the function \code{simulateSPDEmodelMulti}.}

\item{spatialDelta}{a real number greater than zero and less than 1/2 for selecting only the data points which are delta away from the Dirichlet boundary condition. The default is 0.05.}

\item{newColors}{if FALSE, it changes the colors of the returned plots. Default is TRUE.}
}
\value{
a list containing 2 plots and datasets. The main plot displays the theoretical and empirical realized volatilities. The other plot contains its deviation. Also both data sets for creating these plots are provided.
}
\description{
Provides a plot for the realized volatility statistics against the theoretical results. In addition it provides a plot for the deviation of theoretical and empirical results.
}
\examples{
d <- 2
N <- 1000
M <- 10
theta0 <- 0
eta <- 1
nu <- c(2,1)
sigma <- 1
alphaDash <- 0.5
L <- 20
K <- 50
va <- variance_approx(d,theta0,nu,eta,sigma,alphaDash,M,L,K)

res <- simulateSPDEmodelMulti(d=d,theta0=theta0,nu=nu,eta=eta,sigma=sigma,
alphaDash=alphaDash,numberOfSpatialPoints=M,numberOfTemporalPoints=N,
L=L,approx_var=va)

p <- RV_plot(res,0.05,nu,eta,sigma,alphaDash)
p$plot
}
\references{
Bossert, P. (2023), 'Parameter estimation for second-order SPDEs in multiple space dimensions'
}
\seealso{
\link{simulateSPDEmodelMulti},\link{SecondOrderSPDEMulti}
}
\keyword{Plotting}
\keyword{realized}
\keyword{volatilities}
