% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondOrderSPDEMulti.R
\name{SecondOrderSPDEMulti}
\alias{SecondOrderSPDEMulti}
\title{Second-order stochastic partial differential equations in multiple space dimension}
\usage{
SecondOrderSPDEMulti()
}
\description{
This package includes the simulation of SPDE models, \link{simulateSPDEmodelMulti} and \link{MCSPDESamplesMulti}, as well as a quick-plot function \link{plot_SPDEMulti} and a function for estimating the natural parameters of SPDE models \link{estimateParametersMulti}.
}
\seealso{
\link{simulateSPDEmodelMulti}, \link{MCSPDESamplesMulti}, \link{SPDE_plot} and \link{estimateParametersMulti}
}
