% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upsilon.R
\name{Upsilon}
\alias{Upsilon}
\title{Calculate Upsilon}
\usage{
Upsilon(alphaDash, d, n = 10000)
}
\arguments{
\item{d}{natural number which denotes the spacial dimension. 'd' must be greater 1.}

\item{n}{natural number for the approximation of the infinite sum. Default is 10000.}

\item{alphaDsh}{a real number in \code{(0,1)} controlling the roughness of the sample paths.}
}
\value{
A real number of the approximated constant.
}
\description{
Approximation of the constant Upsilon which results by non-negligible covariances of squared increments in the SPDE model.
}
\examples{
Upsilon <- Upsilon(d=2, alphaDash = 1/3)
}
\references{
PhD thesis Bossert, P.
}
\seealso{
\link{SecondOrderSPDEMulti}.
}
\keyword{Approximation}
\keyword{Upsilon}
\keyword{of}
