% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_approx.R
\name{variance_approx}
\alias{variance_approx}
\title{Calculates the approximated variance of the replacement random normals}
\usage{
variance_approx(
  d,
  theta0,
  nu,
  eta,
  sigma,
  alphaDash,
  numberOfSpatialPoints,
  L,
  K,
  numCores = NA
)
}
\arguments{
\item{d}{a natural number, which denotes the spacial dimension. 'd' must be greater 1.}

\item{theta0}{a real number, which controls the drift of the solution field.}

\item{nu}{a real vector of dimension 'd', which controls the curvature of the solution field on each axis, respectively.}

\item{eta}{a real number greater than zero, which reduces the noise level of parameter \code{sigma} and the curvature effect of parameter \code{nu}.}

\item{sigma}{a real number greater than zero, which controls the overall noise level of the solution field.}

\item{alphaDash}{a real number in \code{(0,1)}, controlling the roughness of the temporal marginal process.}

\item{numberOfSpatialPoints}{number of equidistant spatial points M on each axis, where we consider the domain \code{[0,1]}.}

\item{L}{a natural number indicating the replacement bound 'LM', dependent on multiples of the d-dimensional vector with entries \code{M}. The default is \code{L=10}.}

\item{K}{a natural number greater 'L' denoting the cut-off if the variance calculation for the replacement random variables (only of use when method is 'replacement'). 'K' is set to be 'L+10' by default. A too small choice of 'K' produces a bias of the data.}

\item{numCores}{specify the cores to be used for simulation. Default (NA) detects the available cores of the PC and uses all cores minus one.}
}
\value{
A vector containing the approximated variance
}
\description{
Calculates the approximated variance of the replacement random normals within the replacement method. Of use, when considering a Monte carlo sudy. For details, see references or \link{simulateSPDEmodelMulti}.
}
\examples{
d <- 2
M <- 10
theta0 <- 0
eta <- 1
nu <- c(2,1)
sigma <- 1
alphaDash <- 0.5
L <- 20
K <- 50

va <- variance_approx(d,theta0,nu,eta,sigma,alphaDash,M,L,K)
}
\references{
Bossert, P. (2023), 'Parameter estimation for second-order SPDEs in multiple space dimensions'
}
\seealso{
\link{simulateSPDEmodelMulti}
}
\keyword{Approximatex}
\keyword{for}
\keyword{method}
\keyword{replacement}
\keyword{variance}
