% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda.R
\name{Lambda}
\alias{Lambda}
\title{Calculate Upsilon}
\usage{
Lambda(alphaDash, cutoff = 10000)
}
\arguments{
\item{n}{natural number for the approximation of the infinite sum. Default is 10000.}

\item{alphaDsh}{a real number in \code{(0,1)} controlling the roughness of the sample paths.}
}
\value{
A real number of the approximated constant.
}
\description{
Approximation of the constant Lambda which results by non-negligible covariances of squared increments of two temporal resolutions in the SPDE model.
}
\examples{
Lambda <- Lambda(alphaDash = 1/3)
}
\references{
PhD thesis Bossert, P.
}
\seealso{
\link{SecondOrderSPDEMulti}.
}
\keyword{Approximation}
\keyword{Lambda}
\keyword{of}
