% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_plot.R
\name{RV_plot}
\alias{RV_plot}
\title{Plot realized volatility against the theoretical result}
\usage{
RV_plot(data_list, spatialDelta)
}
\arguments{
\item{data_list}{return of the function 'simulateSPDEmodelMulti' or a list containing the returns of the function 'simulateSPDEmodelMulti'.}

\item{spatialDelta}{a real number greater than zero and less than 1/2 for selecting only the data points which are delta away from the Dirichlet boundary condition. The default is 0.05.}

\item{nu}{a real vector of dimension 'd' which controls the curvature of the solution field on each axis respectively.}

\item{eta}{a real number greater than zero which reduces the noise level of parameter sigma and the curvature effect of parameter \code{nu}.}

\item{sigma}{a real number greater than zero which controls the overall noise level of the solution field.}

\item{alphaDash}{a real number in \code{(0,1)} controlling the roughness of the sample paths.}
}
\value{
a list with 2 plots, the main plot displays the theoretical and empirical realized volatilities. The other plot contains its deviation. Also both data sets for creating these plots are provided.
}
\description{
Provides a plot for the realized volatility statistics against the theoretical results. In addition it provides a plot for the deviation of theoretical and empirical results.
}
\examples{
d <- 2
N <- 1000
M <- 10
theta0 <- 0
eta <- 1
nu <- c(2,1)
sigma <- 1
alphaDash <- 0.5
L <- 20
K <- 50
va <- variance_approx(d,theta0,nu,eta,sigma,alphaDash,M,L,K)

res <- simulateSPDEmodelMulti(d=d,theta0=theta0,nu=nu,eta=eta,sigma=sigma,
alphaDash=alphaDash,numberOfSpatialPoints=M,numberOfTemporalPoints=N,
L=L,approx_var=va)

p <- RV_plot(res,0.05,nu,eta,sigma,alphaDash)
p$plot
}
\references{
PhD thesis Bossert, P.
}
\seealso{
\link{simulateSPDEmodelMulti},\link{SecondOrderSPDEMulti}
}
\keyword{Plotting}
\keyword{realized}
\keyword{volatilities}
